requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
asttokens==2.4.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
channels==4.3.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==45.0.6
debugpy==1.8.7
decorator==5.1.1
diff-match-patch==20241021
distlib==0.4.0
Django==5.2.1
django-appconf==1.1.0
django-audit-log==0.7.0
django-auditlog==3.0.0
django-cors-headers==4.7.0
django-decouple==2.1
django-environ==0.11.2
django-extensions==4.1
django-fernet-fields==0.6
django-filter==24.3
django-import-export==4.3.7
django-js-asset==3.0.1
django-model-utils==5.0.0
django-mptt==0.16.0
django-simple-history==3.8.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
docutils==0.21.2
drf-spectacular==0.27.2
drf-yasg==1.21.7
drf-yasg2==1.19.4
drfdocs==0.0.11
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.1.0
Faker==24.11.0
fastapi==0.115.11
fastapi-cli==0.0.7
filelock==3.18.0
h11==0.16.0
httptools==0.6.4
idna==3.10
inflection==0.5.1
ipykernel==6.29.5
ipython==8.28.0
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kombu==5.5.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
nest-asyncio==1.6.0
packaging==24.1
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.0.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
requests==2.32.4
rich==14.1.0
rich-toolkit==0.14.9
rpds-py==0.27.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
tablib==3.8.0
tornado==6.4.1
traitlets==5.14.3
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
virtualenv==20.33.1
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
